backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-PyMongo==2.3.0
pymongo==4.5.0
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.8.0
Pillow==10.0.1
gunicorn==21.2.0
